refactor($ RacingGame, RacingConfig): 패키지 리펙토링
객체를 통제해 비지니스 로직을 관리하는 클래스를 service 패키지로 분리
